Add a basic store for JSON data
Basically I want to be able to easily dump some data into a database and
then access it quickly. This adds a basic version of that, using sqlite
as the store. It is may be faster to use something like lmdb, but I
think this performance will be fast enough for now. I'd also like to use
sled, but that is beta right now.

The design assumes each json object has an id field that is a number to
store. This is a bit different than the current utility that uses a
urs_taxid, but I think it is better.

This also comes with a way query by range of ids. This allows for an
easy check of the number of things extracted.