Rework stores
There is now an LMDB based store that is much faster (<10 seconds vs ~30
for indexing tests on test data). This is now the primary store, but I'm
keeping the sqlite based one around for now. I still need to decide on
the interface for extracting data, as it could be based on ranges
(requires numeric ids) or based upon given ids (could allow urs ids).

I also need to fixup the code to actually use functions and not be so
repetitive.