Try splitting the utilities apart
I want this to be relatively drop in for the existing setup. This means
having a few executable names, which is why I am splitting the data
apart. This creates a few binaries that have names I need, (eg
json2fasta) and gets rid of the old code.

Also, I've added in some old code from other utilities that I think may
be useful here.