Add a utility to extract active xrefs
This may be needed for FTP export. We are having problems with the
database based export. This may be more doable as we can export all
sequences and then the active urs_taxids and then merge the two to
produce the active xref files. This approach builds a set of the ids to
use, it may be possible to use a sorted iterator approach possibly.