Add simple key value store
This is a simple Key-Value CLI store. The idea behind this tool is
provide a way to dump JSON from the database, index it and then pull it
back out for processing it later. It also supports multiple values for a
given key, and will merge all values into a single array. This assumes
all entries have a string 'id' key which is used to identify each entry.
No assumptions about key ordering are made. Attempting to extract a key
that does not exist is an error.

This is needed because the current search export/precompute steps are
very heavily bottlenecked with extracting data from the database. By
using a store like this we can change how the data is extracted. First,
we can split things into smaller joins and then merge things as needed.
Secondly, we can skip grouping things in some cases. Because this can
merge duplicate keys together we can avoid some groupings. These tricks
combined should speed up these steps quite a bit.

This has been tested with some minimal tests to ensure it can handle
basic data, but more through testing on real data will have to happen
now.