Add tool to expand urs -> urs_taxid in JSON
This is useful when I am extracting data for some parts of
search/precompute. In these steps some data is specific to a URS and I
want to avoid doing joins as much as possible. I can do them outside the
database and get faster results, probably. This adds a tool to do that.