Fix some usage
The Reference data structures are copied from python are very
inefficient. They could be coded to have many fewer copies, but I'm not
sure if this is the only place they will be used. The current design
while not efficient does let it be used in most contexts pretty easily.
Once the scope is figured out I should rework them to be better for the
specific use cases.