Rewrite to use RocksDB
Seems that sled is too slow and uses too much memory. This appears to be
a fundamental issue with sled, see:

https://github.com/spacejam/sled/issues/986
https://github.com/the-lean-crate/criner/issues/1

as these discuss the issue pretty clearly. Until the issue is fixed sled
isn't a good choice for our usage. I could switch to sqlite, as I've
written things like that before but I don't want to have. It is harder
than just using a get/merge that is builtin to the database. Rocks has
a similar API but is much faster than sled.

I've explored LMDB before (which is very fast) but it has some sharp
edges that I don't want to deal with, RocksDB while having a weird API
is at least much cleaner to work with I think.